split file in server and programs
